refactor(usage): give each agent its own usage provider
The Usage tab's monitor had grown to four inlined agents in one 1,167-line
file: four credential readers, four endpoint parsers, four log scanners, and
an if-chain plus a copy-pasted enabled/authorized block per agent at the two
call sites. Adding a fifth agent meant editing five places.

Each agent is now a `UsageProvider` in its own file, and `UsageMonitor` only
knows the list — one loop for plan limits, one for the log scan. The shared
plumbing each provider used to restate lives once: the CLI-home lookup, the
JSON credential read/write, the bearer request, the form POST both OAuth
refreshes use, the byte-probed log walk, and a `TokenTally` that owns the
window arithmetic, the day resolver, and the duplicate filter.

Two behavior fixes fall out: no `URL(string:)!` force-unwraps on the endpoint
literals, and one window-naming rule instead of two, so a 7-day Kimi lane
reads "Weekly" rather than "168h".
